Render __webpack_chunks__ in assets.jsx
Resolves #86

Remove OccurrenceOrderPlugin so that server & client chunk IDs are
aligned. With the plugin enabled, the client entry chunk would have ID
N-1 (where N is the total number of chunks), but the server's entry
would have ID 0. This is due to the fact that the client config uses
CommonsChunkPlugin, but the server config does not and cannot:
https://github.com/webpack/webpack/issues/1518
So the server would render the asset path to chunk 1, but the client
would attempt to render the path to chunk 0, causing a React
reconciliation error. Removing OccurrenceOrderPlugin (while keeping
CommonsChunkPlugin) seems to align the chunk IDs for some reason.